requirements.txt
dnspython
pip-chill
psycopg2
pymongo
pytelegrambotapi
python-dotenv
textblob
